mcp_server/requirements.txt
fastapi
uvicorn[standard]
duckdb
pydantic
PyJWT[crypto]
azure-storage-blob
